# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RestReviews API - Restoranlar va Sharhlar",
        "description": "Mobile ilova uchun REST API. Restoranlarni kashf qilish, sharh qoldirish, sevimlilar va boshqalar.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gastronomic.webclub.uz",
            "description": "Production Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://restreviews.test",
            "description": "Laragon Server"
        }
    ],
    "paths": {
...
        "/api/restaurants/{id}/menu": {
            "get": {
                "tags": [
                    "🍔 Menyu"
                ],
                "summary": "Restoran menyusini olish",
                "description": "Restoranning to'liq menyusini menu section lar bilan guruhlanган holda qaytaradi",
                "operationId": "359903fa9a944d987c5d47983ce2e6f4",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (uz, ru, kk, en). Default: kk",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "kk",
                            "enum": [
                                "uz",
                                "ru",
                                "kk",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant menu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/menu-items/{id}": {
            "get": {
                "tags": [
                    "🍔 Menyu"
                ],
                "summary": "Taom batafsil ma'lumotlari",
                "description": "Bitta taomning to'liq ma'lumotlari: nomi, tavsif, narx, rasm, vazn",
                "operationId": "024245ee46acc80d42b2dd4e480a938e",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Til kodi (uz, ru, kk, en). Default: kk",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "kk",
                            "enum": [
                                "uz",
                                "ru",
                                "kk",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu item details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu item not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}